Refactor queue processor base implementation
